New: [U済] バックアップのリムーバブルメディア対応 Patches#609
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@2817 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CEditDoc.h
@@ -369,7 +369,7 @@ protected:
 
 	// 2005.11.21 aroka
 	// 2008.11.23 nasukoji	パスが長すぎる場合への対応
-	bool FormatBackUpPath( char* szNewPath, DWORD dwSize, const char* target_file );	/* バックアップパスの作成 */
+	bool FormatBackUpPath( char* szNewPath, size_t dwSize, const char* target_file );	/* バックアップパスの作成 */
 };
